Reorganize candidate graph module files
