+++ src/motile_toolbox/candidate_graph/__init__.py
@@ -1,3 +1,5 @@
+from .compute_graph import compute_multi_seg_graph, graph_from_segmentation
 from .graph_attributes import EdgeAttr, NodeAttr
-from .graph_from_segmentation import graph_from_segmentation
 from .graph_to_nx import graph_to_nx
+from .iou import add_iou, add_multihypo_iou
+from .utils import add_cand_edges, get_node_id, nodes_from_segmentation
