+++ tests/test_candidate_graph/test_iou.py
@@ -1,15 +1,15 @@
 import pytest
-from motile_toolbox.candidate_graph.iou import compute_ious
+from motile_toolbox.candidate_graph.iou import _compute_ious
 
 
 def test_compute_ious_2d(segmentation_2d):
-    ious = compute_ious(segmentation_2d[0], segmentation_2d[1])
+    ious = _compute_ious(segmentation_2d[0], segmentation_2d[1])
     expected = {1: {2: 555.46 / 1408.0}}
     assert ious.keys() == expected.keys()
     assert ious[1].keys() == expected[1].keys()
     assert ious[1][2] == pytest.approx(expected[1][2], abs=0.1)
 
-    ious = compute_ious(segmentation_2d[1], segmentation_2d[1])
+    ious = _compute_ious(segmentation_2d[1], segmentation_2d[1])
     expected = {1: {1: 1.0}, 2: {2: 1.0}}
     assert ious.keys() == expected.keys()
     assert ious[1].keys() == expected[1].keys()
@@ -19,13 +19,13 @@ def test_compute_ious_2d(segmentation_2d):
 
 
 def test_compute_ious_3d(segmentation_3d):
-    ious = compute_ious(segmentation_3d[0], segmentation_3d[1])
+    ious = _compute_ious(segmentation_3d[0], segmentation_3d[1])
     expected = {1: {2: 0.30}}
     assert ious.keys() == expected.keys()
     assert ious[1].keys() == expected[1].keys()
     assert ious[1][2] == pytest.approx(expected[1][2], abs=0.1)
 
-    ious = compute_ious(segmentation_3d[1], segmentation_3d[1])
+    ious = _compute_ious(segmentation_3d[1], segmentation_3d[1])
     expected = {1: {1: 1.0}, 2: {2: 1.0}}
     assert ious.keys() == expected.keys()
     assert ious[1].keys() == expected[1].keys()
